Relaxed Guided Random Search

rgrs Produce a point configuration given the edge lengths of the desired minimum spanning tree

rgrs(
  edges = NULL,
  d,
  n = NULL,
  theta = NULL,
  outlying = "N",
  skew = "N",
  stringy = "N"
)

Arguments

edges
A numeric vector containing the desired edge lengths of the minimum spanning tree. If n is specified, must be NULL.

d
the dimension of the resulting configuration.

n
the desired number of edge lengths to simulate. If edges is specified, must be set to NULL.

theta
Angle restriction during point proposal of the form (theta1,theta2,p), where p represents the probability of confining the proposal to [theta1,theta2]. Only used for d=2, otherwise NULL. See details for more in depth explanation.

outlying
One of "L", "M", or "H", specifying if the simulated edge lengths should have a Low, Medium, or High outlying scagnostic value.

skew
One of "L", "M", or "H", specifying if the simulated edge lengths should have a Low, Medium, or High skew scagnostic value.

stringy
One of "L", "M", or "H", specifying if the simulated edge lengths should have a Low, Medium, or High stringy scagnostic value. A numeric scalar specifying a value of stringy is also accepted.

Value

An nxd matrix containing the d-dimensional locations of the points.

Details

In 2-dimensions, when a new point is proposed, the position for the new point is determined by:

x <- x0 + r*sin(theta) y <- y0 + r*cos(theta)

where (x0,y0) is the base point, and r is the minimum spanning tree distance. theta is generated from a uniform distribution on (-pi,pi). By specifying the theta argument, the proposed theta is restricted, and is then generated from Uniform(theta1,theta2) or Uniform(-theta2,-theta1) with equal probability. This restriction allows the user to introduce striation into their point configuration.

Examples

# \donttest{
# An example where edge lengths are supplied
EL <- runif(100,0,1)
rgrs(edges = EL, d = 2)

#>               [,1]        [,2]
#>        0.888490296  0.18794620
#>        1.240214597 -0.23471906
#> NewXn  0.893435102  0.50223437
#> NewXn  1.101243302 -1.20072766
#> NewXn  1.115695450 -1.21961863
#> NewXn  1.930673387 -1.23573488
#> NewXn  2.589700081 -1.86506541
#> NewXn  3.093997384 -2.15473972
#> NewXn  0.315757442  0.65747023
#> NewXn  0.286214502  0.73377860
#> NewXn  1.097776083 -1.23852971
#> NewXn  0.117021555 -0.20741116
#> NewXn -0.790263442 -0.23522044
#> NewXn  1.954128004  0.19250938
#> NewXn  4.006335828 -2.53304798
#> NewXn  2.199234546 -2.46301184
#> NewXn  3.409474692 -1.53727822
#> NewXn  0.184571604  0.71646447
#> NewXn -0.216576352  0.98899948
#> NewXn  0.034969077  1.43362811
#> NewXn  2.248779855  0.92993596
#> NewXn  0.771391874  1.33373073
#> NewXn -1.557964045  0.33074410
#> NewXn  2.007042878 -2.33296148
#> NewXn  4.121099590 -2.75094655
#> NewXn  3.337825671 -3.15184887
#> NewXn -1.900216098  0.80816284
#> NewXn  2.441428379 -0.33056691
#> NewXn -0.255818898  0.92645327
#> NewXn  4.052165157 -2.51385454
#> NewXn -2.027428885  1.19732055
#> NewXn  0.495539757  2.11430287
#> NewXn  3.384499141 -0.90978811
#> NewXn  1.966275808 -3.05208217
#> NewXn  2.362561943  0.87353712
#> NewXn -1.166688292 -1.13695161
#> NewXn  0.572844654  2.21569526
#> NewXn  4.967379454 -2.67624208
#> NewXn  2.676788490  0.01504836
#> NewXn -1.414568532  1.34274349
#> NewXn  1.075196781  1.26033599
#> NewXn -2.556577747  1.53992508
#> NewXn  0.564071509  2.24711825
#> NewXn  5.214344225 -2.07801689
#> NewXn  2.268882444  1.30430537
#> NewXn  0.582639825  2.24812708
#> NewXn  3.411630352 -0.80969690
#> NewXn -0.802533874  2.10877190
#> NewXn  1.939356343  1.29264234
#> NewXn  2.576214208  1.32604276
#> NewXn  1.965714341 -2.38806889
#> NewXn  3.853078926 -1.39192172
#> NewXn  4.079325444 -1.29098465
#> NewXn  2.832669628  2.04712743
#> NewXn -1.662984636  2.61002501
#> NewXn  1.539133022 -2.91689206
#> NewXn  1.902984205  2.37033364
#> NewXn  0.659359951 -1.05705815
#> NewXn  5.583185949 -1.48207442
#> NewXn -1.334474721 -1.20080836
#> NewXn  0.665215363  2.38729567
#> NewXn  5.342256736 -1.30764678
#> NewXn  0.663952108 -1.48325843
#> NewXn  3.449630498  1.58285597
#> NewXn  1.960716184 -2.38874867
#> NewXn  3.057105648 -3.81479998
#> NewXn  3.091768760 -4.46117016
#> NewXn  3.126627080 -5.07476296
#> NewXn  3.990355292 -1.15923622
#> NewXn -3.152244613  1.73868329
#> NewXn  0.776240207  3.35002345
#> NewXn  5.425493845 -2.19753645
#> NewXn  3.790514789 -3.82957770
#> NewXn  5.048539935 -0.88669596
#> NewXn -2.665219362  0.33091367
#> NewXn  3.154188668 -5.13429304
#> NewXn -2.931853605  2.21810869
#> NewXn  0.763409612  3.98861813
#> NewXn  5.058830173 -0.59452129
#> NewXn  0.593817073 -3.23459913
#> NewXn  0.189591746  4.28827713
#> NewXn  3.261196459  0.38765570
#> NewXn  5.572261339 -1.01607563
#> NewXn -1.096405398 -1.93512238
#> NewXn  1.252297748  3.19180174
#> NewXn  0.018418782  3.85169771
#> NewXn  2.697472186  1.27924429
#> NewXn  4.030605113  0.63140757
#> NewXn -0.960477905  3.81289890
#> NewXn  3.143455417 -5.25294519
#> NewXn  1.376317926 -2.21228592
#> NewXn  0.441151104 -1.90133464
#> NewXn  0.433241591 -3.97040602
#> NewXn  0.673258604  4.95708639
#> NewXn  3.323196156 -0.88138351
#> NewXn -0.351153650  1.03333903
#> NewXn -1.111448079 -1.94049779
#> NewXn  2.440490218 -4.36194709
#> NewXn -0.008263617  3.33935234
#> NewXn  1.216480273  5.22093644
#> NewXn  0.861305205 -4.52427950

rgrs(edges = EL, d = 3) 

#>               [,1]         [,2]        [,3]
#>        0.347941806  0.983889129  0.35968590
#>       -0.100495515  1.218436378  0.14463367
#> NewXn  0.343940339  0.818727037  0.62709379
#> NewXn -0.822000315  0.571321990  0.25934824
#> NewXn  0.359143280  0.823711159  0.64469386
#> NewXn  0.649218757  1.544154678  0.89221381
#> NewXn  0.567404057  1.028972168 -0.52359035
#> NewXn  1.132997965  1.162893416 -0.54348021
#> NewXn  1.166242478  1.660576310 -0.21330421
#> NewXn -0.825071891  0.492228420  0.23859915
#> NewXn  1.181016777  1.681882498 -0.21585349
#> NewXn  1.772603701  0.679688839 -0.21349643
#> NewXn  1.391632909  2.498101189 -0.55255412
#> NewXn  1.519337588 -0.023524463 -0.57893291
#> NewXn -1.415442981  0.596717532 -0.54627191
#> NewXn  0.007615335  1.763221388  1.11661991
#> NewXn  0.914144060 -0.064828080 -0.91481895
#> NewXn -1.460208175  0.503983461 -0.55151854
#> NewXn  1.013540357  0.862120199 -0.90467319
#> NewXn  0.518797209  0.474668562  0.29240203
#> NewXn -1.205044033  0.391422015  0.93130894
#> NewXn -1.676170106  0.209955255  0.38595762
#> NewXn  1.713826966  0.663628900 -1.52101720
#> NewXn  0.916398911  1.044929053 -1.00953117
#> NewXn  1.429868329  2.741379201 -0.55464461
#> NewXn  0.947007947  2.099457619  1.50638067
#> NewXn  1.463734689  3.278613754 -0.31949098
#> NewXn  0.759307652  3.156799788 -0.32328150
#> NewXn  0.746296534  3.193358711 -0.38609998
#> NewXn -1.218973524  0.434412317  0.95196127
#> NewXn  1.782320420  2.387802119 -0.49941092
#> NewXn  1.885817064  3.622213956 -0.93530658
#> NewXn  2.243335978  2.307913884 -0.08054322
#> NewXn  0.735830215 -0.314627987 -1.46896496
#> NewXn  0.339496991  0.761038574  0.75338315
#> NewXn -0.204369883 -0.257519953  0.32474540
#> NewXn -1.477231454  0.428881221 -0.65313706
#> NewXn -0.825455764  0.655096861  1.67183103
#> NewXn -1.751627448 -0.099862116  0.65644548
#> NewXn  1.442786307 -0.553204805 -0.91108390
#> NewXn  0.786493181 -0.184566182 -0.65587446
#> NewXn -0.003621876  0.388165213  1.12836485
#> NewXn -0.188947872 -0.281226728  0.30848124
#> NewXn  0.134450305  3.387080550 -0.30254751
#> NewXn  0.410676477 -0.268893347 -1.64990683
#> NewXn -0.013026874  0.372376851  1.13120628
#> NewXn  1.956845271  3.611410533 -1.01009049
#> NewXn -0.331471100  1.757298236 -0.64128244
#> NewXn -0.998813037  0.655885033  1.95231296
#> NewXn -0.234946673  1.992277670 -0.81561840
#> NewXn  1.777106450  0.656505533 -1.49475240
#> NewXn -1.041132452  0.454471798  2.37130580
#> NewXn  1.987903439  0.779511602 -0.14238231
#> NewXn  1.749557190  0.499043720  0.52985321
#> NewXn -0.984244035  1.358509240  2.78496049
#> NewXn  1.671527730  0.472270531 -1.92390685
#> NewXn  1.229566179  4.230760525 -0.23370709
#> NewXn  0.720427311 -0.499080854 -0.32028893
#> NewXn  1.447891717  0.674302784 -2.55664815
#> NewXn -0.910300558  1.432980249  2.93061900
#> NewXn -1.082775084  1.485345197  2.80473142
#> NewXn  0.508350371 -0.014072228 -1.53159414
#> NewXn  2.406608232  2.101055468  0.25444861
#> NewXn  1.199826995  1.640140652 -1.41150179
#> NewXn -1.081992002  1.490139769  2.80337402
#> NewXn  1.923077380  1.208467514 -2.64135077
#> NewXn -1.627862760  0.435753144 -1.28262807
#> NewXn -0.172078535  2.320545706  0.93004298
#> NewXn -0.141003317  1.743499740  1.16950555
#> NewXn -1.431898347 -0.002975312 -1.68690413
#> NewXn -2.370957444  0.818525426 -0.65529913
#> NewXn  1.033802827  4.363705630 -0.28729249
#> NewXn  1.923036405 -0.148246659  0.02071615
#> NewXn  2.417464276  3.602670088 -1.23641524
#> NewXn  0.546754132  1.333819649 -1.95248242
#> NewXn -0.195118151 -0.324389709  0.35749507
#> NewXn -1.464867620  0.896868055  0.88799492
#> NewXn  2.573173190  2.590256886  0.62982292
#> NewXn -2.533218694  0.983047124 -0.47620189
#> NewXn  2.934140616  2.836895779 -0.86065562
#> NewXn  2.077354768 -0.493688521 -0.79775631
#> NewXn  0.104112857 -0.436114342 -1.05122094
#> NewXn  2.713844288  1.968284346  0.09353345
#> NewXn  2.276483517  1.863605509 -2.43691889
#> NewXn  1.927068034  0.728470311  0.93914312
#> NewXn  1.537644836  1.604866365  0.07874236
#> NewXn  0.363669368 -0.365709200 -1.72278080
#> NewXn -0.777443485  1.490079814  3.72465508
#> NewXn -0.272432398  2.111464817 -1.78728363
#> NewXn  1.619948162  1.610279114 -0.00722508
#> NewXn -1.601164578  1.484952597  0.97877012
#> NewXn  3.009006753  2.302198051  0.25418452
#> NewXn  0.390236573  2.554471322  1.73035484
#> NewXn -1.321715862  1.648696603  2.02966111
#> NewXn -1.542651311  1.509912545  1.00153140
#> NewXn -0.335836538  1.690988450 -0.76642636
#> NewXn -0.200907254 -0.313557956  0.36770961
#> NewXn  1.699322567  2.482077386  0.24767231
#> NewXn  1.906082996  3.022606646 -0.27479123
#> NewXn  1.770861776  1.311645977  0.95371626
#> NewXn -0.950451112 -0.235715734  2.29760127

 
# An Example where edge lengths are simulated internally
rgrs(d=2, n=100)

#>              [,1]        [,2]
#>        0.07658199  0.69556192
#>        1.01587231  0.79509806
#> NewXn  0.07423902  0.68482333
#> NewXn  1.85790005  1.11129493
#> NewXn -0.32115519  0.42053363
#> NewXn -0.23457853  0.10217171
#> NewXn -0.62590356 -0.72255106
#> NewXn  2.43086807  1.88724338
#> NewXn  2.07009227  2.17765730
#> NewXn  1.10365706  0.51497217
#> NewXn  2.87686537  2.75718233
#> NewXn  2.76090518  1.58300515
#> NewXn  1.14222326  0.44942811
#> NewXn  1.56668230  2.65359907
#> NewXn  2.95697673  1.73974559
#> NewXn  0.70644512  3.09733676
#> NewXn  0.99704673  2.12859107
#> NewXn -0.78377327  0.20948496
#> NewXn  3.08265434  1.37990911
#> NewXn  0.98477181  1.94162249
#> NewXn -0.15666336  0.07015488
#> NewXn -0.29611290  0.96163244
#> NewXn -0.39122159  1.48713215
#> NewXn -0.91975495  0.91663161
#> NewXn  2.12719395  0.72809108
#> NewXn  1.67583434 -0.22549169
#> NewXn -1.66053542  1.18786181
#> NewXn  1.56796369 -0.97236475
#> NewXn  3.31855724  3.00572121
#> NewXn -0.68254229 -0.86256147
#> NewXn -1.54866231 -0.99364107
#> NewXn -0.58399758 -1.07094718
#> NewXn  3.81619955  0.92147374
#> NewXn  3.42045268  2.84989454
#> NewXn -2.10890268  1.60302723
#> NewXn -3.05979102  1.75655690
#> NewXn -0.53874599 -1.11330526
#> NewXn -2.03532220 -0.49713275
#> NewXn  2.07383254 -1.01273458
#> NewXn -2.18369207  0.55235427
#> NewXn  0.12773081 -1.33169247
#> NewXn -0.45988161  1.54612885
#> NewXn  2.17819713 -0.85820800
#> NewXn  3.85812035  0.76981070
#> NewXn  3.48490644  3.11037835
#> NewXn  1.46596610 -1.18737949
#> NewXn -2.86849437  2.62061939
#> NewXn  0.75945623 -0.38865903
#> NewXn -0.33770060  1.90353959
#> NewXn  0.50672191  3.05790238
#> NewXn  4.09810645  1.07269365
#> NewXn  3.36047377  3.47135962
#> NewXn  0.72962319  1.73888022
#> NewXn  1.18070868  0.48846404
#> NewXn -2.25787174 -0.47458654
#> NewXn -3.60037444  1.76649550
#> NewXn -3.03547331  0.33366226
#> NewXn  3.67568116  3.98547255
#> NewXn -2.36331416 -0.58101904
#> NewXn  4.06272957  3.85072297
#> NewXn  0.26074308 -1.99732823
#> NewXn  0.05613376 -2.45850997
#> NewXn  4.02218350  4.57192791
#> NewXn  0.92492228 -2.61026226
#> NewXn  2.87336620 -1.55409648
#> NewXn -4.40297077  1.44112638
#> NewXn  3.83919338  0.41144423
#> NewXn  3.81478411  0.26099215
#> NewXn -3.50446656 -0.17261139
#> NewXn -1.67234237  1.82234559
#> NewXn -1.63086341  2.21404005
#> NewXn  3.93602082  2.71143379
#> NewXn -2.92669094 -1.07809698
#> NewXn -1.05572010  2.67458351
#> NewXn -1.89996424 -0.38061455
#> NewXn  2.48759374  0.40520484
#> NewXn -1.80681330 -1.88818219
#> NewXn -1.70032826 -2.08880245
#> NewXn -3.08335340  0.38010740
#> NewXn  1.29481073  3.81672402
#> NewXn -0.54821599  2.95142008
#> NewXn  1.41298252 -3.31776468
#> NewXn  1.45641108 -1.79442731
#> NewXn -0.71288349 -1.45770198
#> NewXn  4.06341033  3.68880804
#> NewXn -4.46477582 -0.33189626
#> NewXn -3.20012946  0.66426794
#> NewXn -4.92788626  1.63128472
#> NewXn  1.47670627  4.38917246
#> NewXn  4.54792147  2.64281243
#> NewXn -3.83766439  2.55090162
#> NewXn  4.68107811  3.70205067
#> NewXn  4.82099399  4.13495399
#> NewXn  1.81351284 -4.03003531
#> NewXn  2.53137219 -0.62896753
#> NewXn -5.11467416  0.77053983
#> NewXn  3.62204407 -1.49097711
#> NewXn -5.44373599  0.62369567
#> NewXn  4.66056913  1.52781568
#> NewXn -1.04063348  0.17720626
#> NewXn  2.62627872  3.74882470

rgrs(d=3, n=100)

#>              [,1]         [,2]        [,3]
#>        0.75207419  0.141935270  0.86093220
#>        0.97611389 -0.439164824  0.96948647
#> NewXn  0.72353926  0.120612141  0.78694578
#> NewXn  0.50133852  0.520881315  0.69180166
#> NewXn  0.48602147  0.523946038  0.71360954
#> NewXn  0.24679408 -0.879097722  1.16378006
#> NewXn  0.34428452  0.891117236  0.85813888
#> NewXn  0.17855410  1.181173115  0.30467234
#> NewXn  0.50458216  1.582753667  0.06151042
#> NewXn  0.49551451  1.594983204 -0.14412371
#> NewXn  0.78516707  0.065256695  1.49068874
#> NewXn  0.08910622  2.233677091  0.51454663
#> NewXn  0.98907408  0.268045411  0.82376538
#> NewXn  1.57031542 -0.977380858  0.91444956
#> NewXn  1.29539829  2.151713894  0.12550758
#> NewXn  1.09872543  0.404184746  1.79741370
#> NewXn  2.15120775 -1.021029532  1.35977518
#> NewXn  1.51546701 -0.027031036  1.68533661
#> NewXn  0.43602862  0.493876002  0.66407328
#> NewXn  1.59890376 -1.013790476  0.80899271
#> NewXn  0.10078598 -1.178481193  1.01816905
#> NewXn  0.80910487  0.823374793 -0.68293808
#> NewXn  1.38545208  1.434567267 -0.94612113
#> NewXn  2.02443972  1.924892807  0.46467538
#> NewXn  1.75642740  2.975347548  0.35065928
#> NewXn  1.80724967  2.883903374  1.08321551
#> NewXn  0.22497644  0.999231201  0.97499400
#> NewXn  1.80522608  0.360066629  2.37684324
#> NewXn  1.65384622  0.242467951  2.64061057
#> NewXn  0.42506251 -1.185358986  1.37728351
#> NewXn  1.40493250 -0.957364182  1.72986491
#> NewXn  0.42686198  1.006622851  1.40374363
#> NewXn  1.82196854  0.569386854  2.29530509
#> NewXn  1.78988486  1.531001788  1.06242673
#> NewXn  2.25186556  0.688149772  1.31244362
#> NewXn  1.72838522 -0.496411372  0.59125774
#> NewXn -0.07288915 -0.841232235  2.05631293
#> NewXn  1.61482322  2.725302315  1.34076838
#> NewXn  0.91818686  0.618882447  2.30169273
#> NewXn  0.22393504  0.006262072  0.94090753
#> NewXn  0.12512107 -0.067616532  1.48850795
#> NewXn  1.82109053  1.166562147  0.52796606
#> NewXn -0.58314801 -1.513424723  1.91149918
#> NewXn  0.76661281  0.613116071  3.19828430
#> NewXn  2.30158341 -0.375596678  2.69863496
#> NewXn  0.38182989  1.655587425  1.89512115
#> NewXn  1.52319738 -1.096106541  0.82364091
#> NewXn  1.86034676  2.618360650  1.46075413
#> NewXn  0.75363329  0.058114116  0.40313395
#> NewXn  2.06647215  0.604055797  2.49711087
#> NewXn  0.53664527  0.029284654  3.41939596
#> NewXn -0.18223974  2.463885438 -0.19702806
#> NewXn  0.25249980 -1.164722991  0.85396119
#> NewXn  1.52228806 -1.153425689  0.55385619
#> NewXn  0.96120794  0.618152514 -0.60430888
#> NewXn -0.84157398 -1.408125044  1.92426427
#> NewXn  0.22638956  0.054201230  0.35998995
#> NewXn  2.35508137 -0.518808097  2.58023709
#> NewXn  1.84067929 -1.977419024  0.59924160
#> NewXn  1.47304323 -0.099679345  1.75073614
#> NewXn  1.57087319  0.882944925  3.36480885
#> NewXn  1.16869861  1.997737036 -1.05594706
#> NewXn -0.13708324 -0.014045307  0.47112143
#> NewXn  0.81129121  0.491811812 -1.12376836
#> NewXn  2.38920640 -0.259076380  2.81041828
#> NewXn  2.40386930  0.776816564  2.50305987
#> NewXn  2.06921721  0.960918857 -0.05676802
#> NewXn  0.85653523  1.031451709  2.35491038
#> NewXn  2.50310072 -0.626069596  3.12566647
#> NewXn  2.45698279  3.117142745  1.89171119
#> NewXn  2.26276575  0.257282109  0.26241925
#> NewXn  2.27881734  0.262273360  0.28247206
#> NewXn  2.35198579 -0.539726311  2.59500262
#> NewXn  2.41997528 -2.500008193  0.43109880
#> NewXn  0.21082507  0.627942228  1.45342972
#> NewXn  2.79913073  1.117317530  1.39483243
#> NewXn  2.84219734  0.888351387  2.77142538
#> NewXn  0.10080223  1.574237039  1.94692881
#> NewXn  2.38589997  3.392565979  1.92109647
#> NewXn  1.78597036 -2.097271517  1.19018266
#> NewXn  3.39065199  1.260984613  1.92578129
#> NewXn  2.97736063 -0.286104651  0.52555914
#> NewXn  0.16765222 -0.956978944  0.81576829
#> NewXn  1.47892920  1.175984924  0.54123443
#> NewXn -0.28644898  2.436788503 -0.20629668
#> NewXn  3.53005528 -0.034186838 -0.17703360
#> NewXn  1.84436745  3.009486316  0.24750311
#> NewXn  2.88822265 -1.146719754  3.03080237
#> NewXn  2.43385794 -2.572906217  0.38100066
#> NewXn  0.48849452 -0.286825227 -0.07108659
#> NewXn  2.64725727 -2.746370372  0.83172397
#> NewXn  0.33627609  2.538661270 -0.48647808
#> NewXn -0.05364823  1.237971622  0.03394326
#> NewXn  1.53251790  3.255428745  0.12028357
#> NewXn  2.24712010  0.858004255 -0.49150999
#> NewXn  2.96407970  0.382328792  2.13110206
#> NewXn  2.17202565  2.876381971  0.05381506
#> NewXn  1.03697147 -0.307355902  3.74222622
#> NewXn  2.41852757 -1.116464679  1.40794342
#> NewXn  0.81554381  1.043970334  2.31859893
#> NewXn  2.67881438  3.075957216 -0.17540908

rgrs(d=2, n=100, outlying="H")

#>            [,1]      [,2]
#>       0.7038109 0.4246436
#>       0.6953310 0.4301854
#> NewXn 0.7053763 0.4253031
#> NewXn 0.8479552 1.4184697
#> NewXn 0.6750844 0.4230116
#> NewXn 0.6749875 0.4226885
#> NewXn 0.6697798 0.4280413
#> NewXn 0.8489060 1.4182796
#> NewXn 0.8304466 1.4411196
#> NewXn 0.6748410 0.4226414
#> NewXn 0.8536499 1.4210230
#> NewXn 0.8119078 1.4484298
#> NewXn 0.5663337 0.4242783
#> NewXn 0.8119077 1.4484298
#> NewXn 0.7050644 0.4257828
#> NewXn 0.8048607 1.4471616
#> NewXn 0.5663337 0.4242783
#> NewXn 0.6750845 0.4230115
#> NewXn 0.7942350 1.4426558
#> NewXn 0.6953005 0.4301770
#> NewXn 0.5657456 0.4244053
#> NewXn 0.8479554 1.4184694
#> NewXn 0.5661726 0.4233878
#> NewXn 0.5615884 0.4276912
#> NewXn 0.8077643 1.4634954
#> NewXn 0.8078479 1.4639376
#> NewXn 0.6738083 0.4055249
#> NewXn 0.6750415 0.4226894
#> NewXn 0.7052380 0.4259405
#> NewXn 0.8537713 1.4214897
#> NewXn 0.8496159 1.4504132
#> NewXn 0.8406771 1.4777053
#> NewXn 0.8129798 1.4661301
#> NewXn 0.4974115 0.3246712
#> NewXn 0.5501829 0.4403761
#> NewXn 0.8404580 1.4777614
#> NewXn 0.5656779 0.4243229
#> NewXn 0.6729777 0.4311793
#> NewXn 0.7050644 0.4257828
#> NewXn 0.8813443 1.4413757
#> NewXn 0.8095974 1.4417570
#> NewXn 0.7052158 0.4264010
#> NewXn 0.8076003 1.4639664
#> NewXn 0.7926251 1.4600601
#> NewXn 0.8096784 1.4414596
#> NewXn 0.7061888 0.4219683
#> NewXn 0.5502120 0.4404572
#> NewXn 0.5501829 0.4403761
#> NewXn 0.8098021 1.4413441
#> NewXn 0.6538677 0.3268152
#> NewXn 0.8540987 1.4210102
#> NewXn 0.8078471 1.4639374
#> NewXn 0.6388157 0.3185916
#> NewXn 0.7925678 1.4600707
#> NewXn 1.0265311 1.6406923
#> NewXn 0.8070792 1.4392617
#> NewXn 0.6500901 0.3007042
#> NewXn 0.7864577 1.4408073
#> NewXn 0.8399969 1.4791070
#> NewXn 0.4962869 0.3235572
#> NewXn 1.1246692 1.6866752
#> NewXn 1.1506403 1.6929299
#> NewXn 0.7925707 1.4600810
#> NewXn 0.8098335 1.4413009
#> NewXn 0.7804102 1.4396271
#> NewXn 0.8813455 1.4413757
#> NewXn 1.1524987 1.6642102
#> NewXn 1.2590947 1.7474421
#> NewXn 1.1521839 1.6550806
#> NewXn 0.8098927 1.4412200
#> NewXn 0.7636859 1.4419461
#> NewXn 0.5468463 0.4409223
#> NewXn 1.2960746 1.7729921
#> NewXn 0.8404564 1.4777649
#> NewXn 0.7099044 0.4200031
#> NewXn 0.7054638 0.4259980
#> NewXn 0.7925752 1.4600295
#> NewXn 0.6501233 0.3006717
#> NewXn 0.7050982 0.4271031
#> NewXn 1.1200799 1.6980234
#> NewXn 0.7119505 0.4205360
#> NewXn 1.0422856 2.1933616
#> NewXn 1.1135190 2.2556741
#> NewXn 0.8304466 1.4411196
#> NewXn 0.6296151 0.2469423
#> NewXn 0.4973085 0.3219038
#> NewXn 0.4830129 0.3195720
#> NewXn 0.6385078 0.3185566
#> NewXn 0.4828604 0.3198737
#> NewXn 1.1209563 2.2606055
#> NewXn 0.8075152 1.4645956
#> NewXn 0.7890655 1.4631048
#> NewXn 0.5468502 0.4411837
#> NewXn 0.6384111 0.3193465
#> NewXn 1.1521898 1.6550768
#> NewXn 1.1520035 1.6544825
#> NewXn 0.5468286 0.4409230
#> NewXn 0.4969234 0.4669108
#> NewXn 1.1219319 2.2608263
#> NewXn 1.1637714 1.5414734
#> NewXn 1.3086231 1.7671778

rgrs(d=2, n=100, skew = "M")

#>               [,1]         [,2]
#>         0.45056883   0.08929805
#>         0.03141547   1.86026246
#> NewXn   0.66498012  -0.76994004
#> NewXn   2.21854659  -1.41121590
#> NewXn   2.11626276  -1.84924023
#> NewXn   2.64171233  -2.48743108
#> NewXn   0.31343889  -2.36877561
#> NewXn  -0.37492459  -0.65524915
#> NewXn  -1.63367632   2.99394360
#> NewXn  -2.97010212   3.31403257
#> NewXn   3.36447024   0.51927966
#> NewXn   2.31632684  -4.95323234
#> NewXn  -1.56920370   4.03652412
#> NewXn   3.54991857   1.06665363
#> NewXn   2.89547368  -5.27711986
#> NewXn   2.24488186   3.83154730
#> NewXn  -3.98095021   3.76330160
#> NewXn   4.21390410   0.44212952
#> NewXn  -4.21615121   2.58379085
#> NewXn   1.54934444  -5.08030863
#> NewXn  -4.90313169   1.68129173
#> NewXn   2.33528282   4.96505052
#> NewXn  -5.59705032   3.82317675
#> NewXn  -6.19740164   3.13123065
#> NewXn   4.46703064   0.79339171
#> NewXn   5.28412465   0.03689850
#> NewXn   2.15939600   5.78448878
#> NewXn   1.19037627  -5.50948125
#> NewXn   0.46772682  -6.76697863
#> NewXn   3.55579668   4.33413809
#> NewXn   6.60313680  -0.18586401
#> NewXn  -0.08201461  -9.08889030
#> NewXn   0.98724910 -10.22723306
#> NewXn   4.81033216   3.86732641
#> NewXn   8.45829753   2.43895192
#> NewXn   6.53777032  -0.68981141
#> NewXn   3.63960926  -2.67713868
#> NewXn  -0.30376429 -10.90580844
#> NewXn   1.50746670   6.27299724
#> NewXn  -0.39495092  -2.43314301
#> NewXn  -5.84579158   0.63641616
#> NewXn  -1.05013250  -0.65560483
#> NewXn  -6.46448681   0.31398574
#> NewXn  -0.74545404  -3.09355242
#> NewXn   2.16400432 -11.68258475
#> NewXn   3.89718571  -2.63302807
#> NewXn  -1.48179073   9.02695849
#> NewXn   3.94536831 -10.50955851
#> NewXn   5.44737448 -10.61194441
#> NewXn  -1.46023715  10.48429783
#> NewXn   2.05953168  -5.80032599
#> NewXn  -6.39509755   1.45788896
#> NewXn  -7.65405709   0.85509960
#> NewXn  -7.16093759  -1.88187119
#> NewXn  11.73523840   2.28033928
#> NewXn  -0.39540129 -12.44181169
#> NewXn  -1.39565226  -2.55518090
#> NewXn  -2.72107898  -7.34572999
#> NewXn  -0.22614858 -13.99896633
#> NewXn  -1.75902993 -14.20935183
#> NewXn  -3.39130627   4.98818252
#> NewXn   3.17216563  -5.97760408
#> NewXn   5.49307671   4.08447214
#> NewXn   4.98084912  -5.02462832
#> NewXn  -4.06354217   0.44183082
#> NewXn   5.42794109   4.24966023
#> NewXn  12.14033803   3.56705961
#> NewXn  -6.65980530  -3.13686063
#> NewXn  12.55974873   4.94163948
#> NewXn  -2.30137916   1.51776084
#> NewXn  10.30577608   5.57758910
#> NewXn   6.96490578  -6.88744342
#> NewXn  -1.54992883 -16.37920861
#> NewXn  13.57892787   5.71023148
#> NewXn  -9.10946639  -1.51884858
#> NewXn  13.93197021   5.37448804
#> NewXn -10.73223143   0.04820642
#> NewXn   2.33673099  -7.79368033
#> NewXn   7.69165736  -8.98365292
#> NewXn   7.81230168  -6.69311594
#> NewXn  -1.45083320 -17.43334079
#> NewXn   6.47834105  -1.37582305
#> NewXn   4.18906690   8.24662119
#> NewXn -10.03820595  -4.43475045
#> NewXn  -2.47685021 -13.97988282
#> NewXn  12.77236812   1.64081350
#> NewXn  -9.62679500  -5.08384919
#> NewXn  -4.79720773  -2.29976043
#> NewXn  -3.48563451 -13.00167068
#> NewXn  14.58547001   5.85013302
#> NewXn  -2.71994643  -6.86497830
#> NewXn  -1.82575062  -0.74961368
#> NewXn  12.71131527  -0.08027650
#> NewXn   0.25125835 -13.97778608
#> NewXn  15.43095731   5.16989447
#> NewXn   7.06697773  -2.21081632
#> NewXn  -6.32680295   4.32765490
#> NewXn  -1.48389874  11.57875912
#> NewXn   5.42334705   9.40922336
#> NewXn  -3.78786484  11.03618783
#> NewXn   0.64073653 -13.65842981

rgrs(d=2, n=100, stringy = "H")

#>              [,1]        [,2]
#>        0.47088818  0.68210089
#>        0.76648398  0.84996401
#> NewXn  1.13235024  0.92045436
#> NewXn -0.03884172  0.75906402
#> NewXn -0.22248915  0.25020619
#> NewXn  1.14507420  0.78982924
#> NewXn -0.23117911  0.25325809
#> NewXn  1.84400594  0.40370613
#> NewXn  1.98341892  0.29821710
#> NewXn -0.42525108 -0.08579356
#> NewXn  1.80023562 -0.06091845
#> NewXn  1.95828006 -0.11596219
#> NewXn -0.50633651 -0.69016710
#> NewXn  2.33054802 -0.95033827
#> NewXn -1.10324443 -1.45226273
#> NewXn -1.67749096 -1.40290013
#> NewXn  2.31415092 -1.01339763
#> NewXn -1.88510031 -1.72563733
#> NewXn -2.10929722 -2.34440077
#> NewXn  1.91002961 -1.39899265
#> NewXn -2.64404403 -3.10947630
#> NewXn -2.55933251 -3.95511554
#> NewXn -1.95911018 -4.72643042
#> NewXn  1.64617103 -1.75808186
#> NewXn -2.67594713 -5.23219996
#> NewXn  1.49099050 -1.76040225
#> NewXn -2.73566989 -5.29441172
#> NewXn  1.26998250 -2.65720666
#> NewXn -2.49373122 -5.55544273
#> NewXn -2.40659493 -5.77914495
#> NewXn  1.46701037 -3.22379959
#> NewXn  0.79123319 -3.68477314
#> NewXn -1.53256359 -5.73421893
#> NewXn -1.11570238 -5.13876634
#> NewXn  1.42482780 -4.34135840
#> NewXn -0.31852837 -5.41050769
#> NewXn  1.99223316 -4.03579488
#> NewXn  0.48913682 -5.38264100
#> NewXn  2.55456943 -4.11757592
#> NewXn  0.70862674 -5.44399898
#> NewXn  0.76425195 -5.23276543
#> NewXn  2.80242236 -3.25799906
#> NewXn  2.85701794 -3.26019521
#> NewXn  3.52918050 -2.56310524
#> NewXn  3.30813734 -2.38874691
#> NewXn  0.68276242 -5.02390760
#> NewXn  3.17140715 -2.15615698
#> NewXn  3.39322424 -1.55982004
#> NewXn  0.27657614 -4.72528707
#> NewXn  3.88931822 -1.35403392
#> NewXn  3.43183202 -0.67658701
#> NewXn  3.27919835 -0.35211824
#> NewXn  3.14917632  0.46242039
#> NewXn -0.03113235 -4.36554828
#> NewXn  3.98836539  0.89746587
#> NewXn -0.87257135 -4.21324645
#> NewXn  4.88617050  0.95772902
#> NewXn  5.49534321  0.91318387
#> NewXn -1.38537227 -3.62535708
#> NewXn  6.32428038  1.32557317
#> NewXn  6.60715090  1.83977275
#> NewXn  6.76115024  2.60708724
#> NewXn  7.02978814  3.37613684
#> NewXn  7.51588200  3.49408216
#> NewXn -1.43357738 -3.32025236
#> NewXn -1.00816788 -3.37575054
#> NewXn -0.66048432 -2.85191786
#> NewXn -0.57585301 -2.87098240
#> NewXn  7.84882508  2.93958242
#> NewXn  7.90108757  2.31953860
#> NewXn -0.36613823 -2.21029953
#> NewXn  8.09762706  1.41598394
#> NewXn  8.08754874  1.21181391
#> NewXn  7.29244137  1.03404046
#> NewXn  7.22886678  0.80394566
#> NewXn  0.48371421 -1.73730228
#> NewXn  6.57540068  0.39069422
#> NewXn  0.84524433 -0.82701186
#> NewXn  5.96020953  0.08744139
#> NewXn  0.93321876 -0.72752480
#> NewXn  5.61346754 -0.18371965
#> NewXn  6.19185862 -0.57366575
#> NewXn  0.79048033 -0.46908974
#> NewXn  0.74573867 -0.34906686
#> NewXn  6.84835489 -0.66697634
#> NewXn  7.02586282 -0.33341756
#> NewXn  7.62238151 -0.81275170
#> NewXn  8.34446538 -1.04241066
#> NewXn  8.37223752 -1.05567523
#> NewXn  0.83464579 -0.25425934
#> NewXn  8.37719142 -1.05491018
#> NewXn  8.44863560 -1.23526937
#> NewXn  8.80465813 -2.13577912
#> NewXn  9.21476638 -2.49862299
#> NewXn  8.59300401 -2.98349972
#> NewXn  8.81019337 -3.30817613
#> NewXn  8.65172034 -3.54690950
#> NewXn  8.43256830 -3.71541878
#> NewXn  0.72427884 -5.48956876
#> NewXn -1.99799452 -0.89896878
#> NewXn -2.93648508 -5.21744885

 
# An Example making use of theta
rgrs(d=2, n=100, theta=c(pi/4,pi/3,.5))

#>              [,1]        [,2]
#>        0.57857482  0.87069059
#>        0.50167305  0.77608673
#> NewXn  0.93265119  1.22954200
#> NewXn  1.15453274  1.54702452
#> NewXn  1.52570057  2.24862574
#> NewXn  0.38378672  0.62116999
#> NewXn  0.04519500  0.17087541
#> NewXn  1.80101885  1.12727087
#> NewXn  1.39519198  2.57027481
#> NewXn  0.68053843  2.10580838
#> NewXn  1.87542061  1.20941907
#> NewXn  0.18550150  1.39508161
#> NewXn  1.59873191  2.82244910
#> NewXn -0.38141307 -0.36365835
#> NewXn -0.40536941 -0.34991314
#> NewXn -0.64558182 -0.66803632
#> NewXn  1.58393079  0.28818931
#> NewXn  1.06334014 -0.51097876
#> NewXn  0.63207954 -1.14546159
#> NewXn  1.41566166  2.90697292
#> NewXn  2.33552550  1.86391219
#> NewXn  2.45761874  2.01521852
#> NewXn -0.84753624 -0.94096748
#> NewXn  0.51570132  2.16960377
#> NewXn -1.81813086 -0.77324733
#> NewXn -1.74031785  0.07222848
#> NewXn -1.22241640  0.62371219
#> NewXn -1.19996261  0.65129071
#> NewXn -1.24128560  0.69298567
#> NewXn  1.40770426 -1.24983371
#> NewXn  1.61416436  3.24221792
#> NewXn -1.95286907 -0.94690271
#> NewXn -0.82697383  1.39501033
#> NewXn -2.17503408 -1.19629803
#> NewXn -0.14800220  2.58183944
#> NewXn  3.35663265  1.91118393
#> NewXn  2.22606360 -0.97609070
#> NewXn  0.02247438 -1.84332519
#> NewXn  1.40909505  2.90034111
#> NewXn  1.20132032  3.04856639
#> NewXn -2.53940532 -1.04794228
#> NewXn  1.08617483  3.10060568
#> NewXn  1.66387936  3.21527972
#> NewXn -0.34279830  2.28010517
#> NewXn  2.35295161 -0.82589419
#> NewXn  3.84188664  2.66594627
#> NewXn  2.42281956 -0.71029249
#> NewXn -2.91851313 -1.68422294
#> NewXn -2.87672109 -1.82067502
#> NewXn  1.04624026  3.74401676
#> NewXn  2.15257697 -0.49344447
#> NewXn  2.75116653  1.01519007
#> NewXn -0.70988168  2.74365069
#> NewXn -2.35096190  0.03571481
#> NewXn -0.49385814 -2.60419717
#> NewXn -3.18446510 -1.94368610
#> NewXn  2.93935393  2.73237215
#> NewXn  1.29173746  4.28544233
#> NewXn  3.96229477  2.80985818
#> NewXn  1.51354349 -1.56993789
#> NewXn -3.11819677 -2.21863775
#> NewXn -1.05267692  2.71484175
#> NewXn -3.53712424 -2.63297284
#> NewXn -0.79689621  3.18886651
#> NewXn -1.51625920  3.81071627
#> NewXn  2.51149079 -0.04263143
#> NewXn  1.67047071  4.72812052
#> NewXn -1.87892257  3.43386188
#> NewXn  2.86051515  1.13385362
#> NewXn -0.85933048 -3.15178499
#> NewXn -3.05164138 -3.27023331
#> NewXn -0.25953237  3.76318913
#> NewXn -1.07361556 -2.16853534
#> NewXn -1.19544790  4.35982945
#> NewXn  4.11136283  3.06554507
#> NewXn  1.21848449 -1.87856363
#> NewXn  1.59789593  2.28372485
#> NewXn  2.61494282 -0.92003703
#> NewXn -0.88038370  4.90308119
#> NewXn  0.91227845  4.61391906
#> NewXn  1.01698862  4.75498823
#> NewXn  0.75822881 -2.54000775
#> NewXn -1.58914212 -2.69396870
#> NewXn  1.46021129 -3.06816012
#> NewXn  1.54216271 -1.59590439
#> NewXn -0.85963252  4.93638559
#> NewXn -2.04898196  0.75847870
#> NewXn -0.49860394  5.48400306
#> NewXn -1.90550225  3.32422622
#> NewXn  1.58126042 -2.17613834
#> NewXn -2.68174237 -2.90949861
#> NewXn -2.11436537  4.43366414
#> NewXn -0.27481719  5.52923076
#> NewXn  1.59444488  3.29481274
#> NewXn  4.25962390  2.99090909
#> NewXn  2.01625361  5.09175367
#> NewXn -1.06524569 -3.47964930
#> NewXn  0.05991361  4.30818622
#> NewXn  1.04576480 -3.70696961
#> NewXn -1.73847868 -3.48187247
#> NewXn -0.46162702 -4.08018049

# }